+++ mlrun/runtimes/sparkjob/spark3job.py
@@ -20,7 +20,7 @@ import mlrun.common.schemas.function
 import mlrun.errors
 import mlrun.runtimes.pod
 
-from ...utils import update_in, verify_and_update_in, verify_field_regex
+from ...utils import update_in, verify_field_regex
 from ..utils import (
     generate_resources,
     get_gpu_from_resource_requirement,
@@ -388,116 +388,13 @@ class Spark3JobSpec(AbstractSparkJobSpec):
 
 
 class Spark3Runtime(AbstractSparkRuntime):
-    def _enrich_job(self, job):
-        if self.spec.priority_class_name:
-            verify_and_update_in(
-                job,
-                "spec.batchSchedulerOptions.priorityClassName",
-                self.spec.priority_class_name,
-                str,
-            )
+    @property
+    def spec(self) -> Spark3JobSpec:
+        return self._spec
 
-        verify_and_update_in(
-            job,
-            "spec.driver.cores",
-            self.spec.driver_cores or 1,
-            int,
-        )
-        # By default we set this to 1 in the parent class. Here we override the value if requested.
-        if self.spec.executor_cores:
-            verify_and_update_in(
-                job,
-                "spec.executor.cores",
-                self.spec.executor_cores,
-                int,
-            )
-
-        if self.spec.dynamic_allocation:
-            if "enabled" in self.spec.dynamic_allocation:
-                update_in(
-                    job,
-                    "spec.dynamicAllocation.enabled",
-                    self.spec.dynamic_allocation["enabled"],
-                )
-            if "initialExecutors" in self.spec.dynamic_allocation:
-                update_in(
-                    job,
-                    "spec.dynamicAllocation.initialExecutors",
-                    self.spec.dynamic_allocation["initialExecutors"],
-                )
-            if "minExecutors" in self.spec.dynamic_allocation:
-                update_in(
-                    job,
-                    "spec.dynamicAllocation.minExecutors",
-                    self.spec.dynamic_allocation["minExecutors"],
-                )
-            if "maxExecutors" in self.spec.dynamic_allocation:
-                update_in(
-                    job,
-                    "spec.dynamicAllocation.maxExecutors",
-                    self.spec.dynamic_allocation["maxExecutors"],
-                )
-        update_in(job, "spec.driver.serviceAccount", "sparkapp")
-        update_in(
-            job, "spec.executor.serviceAccount", self.spec.service_account or "sparkapp"
-        )
-        if self.spec.driver_node_selector:
-            update_in(job, "spec.driver.nodeSelector", self.spec.driver_node_selector)
-        if self.spec.executor_node_selector:
-            update_in(
-                job, "spec.executor.nodeSelector", self.spec.executor_node_selector
-            )
-        if self.spec.driver_tolerations:
-            update_in(job, "spec.driver.tolerations", self.spec.driver_tolerations)
-        if self.spec.executor_tolerations:
-            update_in(job, "spec.executor.tolerations", self.spec.executor_tolerations)
-
-        if self.spec.driver_affinity:
-            update_in(job, "spec.driver.affinity", self.spec.driver_affinity)
-        if self.spec.executor_affinity:
-            update_in(job, "spec.executor.affinity", self.spec.executor_affinity)
-
-        if self.spec.monitoring:
-            if "enabled" in self.spec.monitoring and self.spec.monitoring["enabled"]:
-                update_in(job, "spec.monitoring.exposeDriverMetrics", True)
-                update_in(job, "spec.monitoring.exposeExecutorMetrics", True)
-                if "exporter_jar" in self.spec.monitoring:
-                    update_in(
-                        job,
-                        "spec.monitoring.prometheus.jmxExporterJar",
-                        self.spec.monitoring["exporter_jar"],
-                    )
-
-        if self.spec.driver_volume_mounts:
-            update_in(
-                job,
-                "spec.driver.volumeMounts",
-                self.spec.driver_volume_mounts,
-                append=True,
-            )
-        if self.spec.executor_volume_mounts:
-            update_in(
-                job,
-                "spec.executor.volumeMounts",
-                self.spec.executor_volume_mounts,
-                append=True,
-            )
-        if self.spec.driver_java_options:
-            update_in(
-                job,
-                "spec.driver.javaOptions",
-                self.spec.driver_java_options,
-            )
-        if self.spec.executor_java_options:
-            update_in(
-                job,
-                "spec.executor.javaOptions",
-                self.spec.executor_java_options,
-            )
-        return
-
-    def _get_spark_version(self):
-        return "3.1.2"
+    @spec.setter
+    def spec(self, spec):
+        self._spec = self._verify_dict(spec, "spec", Spark3JobSpec)
 
     def _get_igz_deps(self):
         return {
@@ -510,14 +407,6 @@ class Spark3Runtime(AbstractSparkRuntime):
             ],
             "files": ["local:///igz/java/libs/v3io-pyspark.zip"],
         }
-
-    @property
-    def spec(self) -> Spark3JobSpec:
-        return self._spec
-
-    @spec.setter
-    def spec(self, spec):
-        self._spec = self._verify_dict(spec, "spec", Spark3JobSpec)
 
     def with_node_selection(
         self,
